call smoke tests feature from RunCucumberTestCases runner so that they get called first
