+++ src/test/java/uk/gov/ons/ctp/integration/contcencucumber/RunCucumberTestCases.java
@@ -7,7 +7,7 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(
     plugin = {"pretty", "html:target/cucumber"},
-    features = {"src/test/resources/integrationtests/case"},
+    features = {"src/test/resources/integrationtests/case/smokeTests.feature", "src/test/resources/integrationtests/case/testCaseEndpoints.feature"},
     glue = {"uk.gov.ons.ctp.integration.contcencucumber.cucSteps.cases"},
     dryRun = false)
 public class RunCucumberTestCases {}
