+++ src/test/resources/integrationtests/case/smokeTests.feature
